feat: refactor post model to utilize utility functions for media, likes, comments, and liked status
